Add ModificationHealthRegen; Boss heals over time based on its current BossModification's healthRegeneration mod. Add AccumulationHealing; This accumulation strategy adds to the healthRegeneration mod.

+++ Assets/Scripts/Player/PlayerAbilities/AbilityShoot/AbilityShootCollision.cs
@@ -5,7 +5,7 @@ using Mirror;
 
 public class AbilityShootCollision : NetworkBehaviour
 {
-    int DAMAGE = 10;
+    float DAMAGE = 10;
 
     void OnTriggerEnter2D(Collider2D collider)
     {
